permissions for all users

+++ resources/views/users/index.blade.php
@@ -6,7 +6,9 @@
             <h2>Users Management</h2>
         </div>
         <div class="float-end">
+            @can('user-create')
             <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+            @endcan
         </div>
     </div>
 </div>
@@ -42,10 +44,14 @@
                 </td>
                 <td>
                     <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
+                    @can('user-edit')
                     <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+                    @endcan
+                    @can('user-delete')
                         {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
                             {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                         {!! Form::close() !!}
+                    @endcan
                 </td>
             </tr>
             @endforeach
